Refactoring * tests/auth_spawn_rust rename to auth-spawn-tests * tests/cardano_lock move to auth-spawn-tests * Add Makefile in tests for testing all * All install for testcase move to tests/Makefile
