+++ .github/workflows/rust.yml
@@ -12,36 +12,14 @@ jobs:
     - uses: actions/checkout@v3
       with:
         submodules: recursive
-    - name: Install litecoin for tests
-      run: |
-        tarball=litecoin.tar.gz
-        wget -O "$tarball" https://download.litecoin.org/litecoin-0.21.2.2/linux/litecoin-0.21.2.2-x86_64-linux-gnu.tar.gz
-        tar xvzf "$tarball"
-        sudo cp -r litecoin-*/* /usr/local/
-        litecoin-cli --help
-    - name: Install monero wallet cli for tests
-      run: |
-        tarball=monero-wallet-cli.tar.gz
-        wget -O "$tarball" https://downloads.getmonero.org/cli/monero-linux-x64-v0.18.2.2.tar.bz2
-        tar xvaf "$tarball"
-        sudo cp -r monero-*/* /usr/local/bin/
-        monero-wallet-cli --help
-    - name: Install solana cli for tests
-      run: |
-        tarball=solana-release-x86_64-unknown-linux-gnu.tar.bz2
-        wget -O "$tarball" https://github.com/solana-labs/solana/releases/download/v1.16.1/solana-release-x86_64-unknown-linux-gnu.tar.bz2
-        tar xvaf "$tarball"
-        sudo cp -r solana-*/* /usr/local/
-        solana --help
-        solana-keygen new --force --no-bip39-passphrase
+    - name: Install tools
+      run: cd tests && make install_all
     - name: Build contract
       run: make all-via-docker
     - name: Run auth_rust tests
       run: cd tests/auth_rust && bash run.sh
     - name: Clean auth_rust
       run: rm -rf tests/auth_rust/target
-    - name: Install ckb-debugger
-      run: cd tests/auth_spawn_rust && make install
     - name: Run auth_spawn_rust tests
       run: cd tests/auth_spawn_rust && make all
     - name: Install cardano tools
