+++ tests/auth-spawn-tests/.gitignore
