+++ tests/auth-spawn-tests/src/auto_complete.rs
