+++ tests/auth-spawn-tests/src/bin/auth-spawn-rust-success.rs
