+++ tests/auth-spawn-tests/src/bin/auth-spawn-success.rs
