+++ tests/auth-spawn-tests/src/bin/ripple-success.rs
