+++ tests/auth-spawn-tests/src/hash.rs
