+++ tests/auth-spawn-tests/src/smt.rs
