+++ tests/auth-spawn-tests/templates/auth-spawn-success.json
