+++ tests/auth-spawn-tests/templates/cardano-success.json
