+++ tests/auth-spawn-tests/templates/ripple-success.json
