--- tests/cardano_lock/.gitignore
@@ -1,3 +1,0 @@
-tx.json
-/bin/
-test_data/
