--- tests/cardano_lock/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "cardano_lock"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-auth_spawn_rust = {path = "../auth_spawn_rust"}
-ckb-debugger-api = { git = "https://github.com/nervosnetwork/ckb-standalone-debugger.git", rev="5077c6c" }
-ckb-mock-tx-types =  { git = "https://github.com/nervosnetwork/ckb-standalone-debugger.git", rev="5077c6c" }
-molecule = { version = "0.7.3", default-features = false }
-ckb-hash = "=0.109.0"
-ckb-jsonrpc-types = "=0.109.0"
-ckb-types = "=0.109.0"
-ckb-crypto = "=0.109.0"
-env_logger = "0.4.3"
-lazy_static = "1.4.0"
-log = "0.4.0"
-serde_json = "1.0"
-serde_plain = "1.0"
-sparse-merkle-tree = {version = "0.6.1"}
-cardano-serialization-lib = "11.4.0"
-cardano-message-signing = { git = "https://github.com/Emurgo/message-signing.git" }
-hex = "0.4.3"
-cbor_event = "2.4.0"
-blake2b-rs = "0.2.0"
